requirements.txt
feedparser==6.0.11
python-dotenv==1.0.0
